refactor: 코리반영 (커스텀훅 분리)
